+++ src/app/filters/ActionFilterBase.ts
@@ -1,3 +1,7 @@
+import { injectable } from '@micro-fleet/common'
+
+
+@injectable()
 export abstract class ActionFilterBase {
 
     protected addReadonlyProp(obj: object, prop: string, value: any): void {