+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## VERSIONS
 
+### 1.1.2
+- Added `request.extras` property as attachment point for filters.
+- [#4](https://github.com/gennovative/micro-fleet-web/issues/4) - Global ErrorHandlerFilter cannot catch async exception from action method
+- [#5](https://github.com/gennovative/micro-fleet-web/issues/5) - Should allow route param in URL prefix setting
+
 ### 1.1.1
 - Remove script "postinstall" from `package.json`.